+++ pytket/tests/predicates_test.py
@@ -1107,6 +1107,19 @@ def test_auto_rebase_deprecation(recwarn: Any) -> None:
     assert "deprecated" in str(w.message)
 
 
+def test_get_pre_conditions() -> None:
+    pre_cons = GreedyPauliSimp().get_pre_conditions()
+    gate_set = pre_cons[0].gate_set
+    assert OpType.CX in gate_set
+    assert OpType.Measure in gate_set
+
+
+def test_get_post_conditions() -> None:
+    gate_set = {OpType.CX, OpType.Rz, OpType.H, OpType.Reset, OpType.Measure}
+    post_cons = AutoRebase(gate_set).get_post_conditions()
+    assert post_cons[0].gate_set == gate_set
+
+
 if __name__ == "__main__":
     test_predicate_generation()
     test_compilation_unit_generation()